Refactor CPP Modules 01-04 to meet subject requirements and best practices
- CPP01: Added destructor messages, fixed output formats, added NULL checks.
- CPP02: Added setRawBits, fixed C++98 compatibility, improved arithmetic precision.
- CPP03: Fixed encapsulation, added virtual destructors, fixed assignment operators.
- CPP04: Implemented deep copy for Brain class, fixed memory management in Cat/Dog.

+++ cpp02/ex01/Fixed.cpp
@@ -13,8 +13,8 @@ Fixed::Fixed(const int n)
 
 Fixed::Fixed(const float f)
 {
-	std::cout << "Float constructor called" << std::endl;
-	this->value = roundf(f * (1 << bits));
+        std::cout << "Float constructor called" << std::endl;
+        this->value = (int)(round(f * (1 << bits)));
 }
 
 Fixed::Fixed(const Fixed& other)
